Provide legacy implementation for parsing Format string
