+++ Sources/RichStringKit/Primitives/Format.swift
@@ -1,4 +1,3 @@
-@available(iOS 16, tvOS 16, watchOS 9, macOS 13, *)
 public struct Format: RichString {
     public typealias Body = Never
 
@@ -18,7 +17,6 @@ public struct Format: RichString {
     }
 }
 
-@available(iOS 16, tvOS 16, watchOS 9, macOS 13, *)
 extension Format {
     public init(
         _ formatString: String,
